refactor(ghost): consolidate types into types.ts
Move all type definitions from individual files to types.ts for better organization:
- GhostStatusBarStateProps from GhostStatusBar.ts
- AutocompleteContext, CacheMatchType from AutocompleteTelemetry.ts
- FimGhostPrompt, FimCompletionResult from FillInTheMiddle.ts
- HoleFillerGhostPrompt, FillInAtCursorSuggestion, ChatCompletionResult from HoleFiller.ts
- CostTrackingCallback, GhostPrompt, MatchingSuggestionResult, LLMRetrievalResult, PendingRequest from GhostInlineCompletionProvider.ts
- AUTOCOMPLETE_PROVIDER_MODELS, AutocompleteProviderKey from kilocode-utils.ts

All 159 tests pass.
