+++ src/services/ghost/classic-auto-complete/FillInTheMiddle.ts
@@ -1,19 +1,15 @@
-import { AutocompleteInput, ResponseMetaData } from "../types"
-import { GhostContextProvider } from "./GhostContextProvider"
+import {
+	AutocompleteInput,
+	GhostContextProvider,
+	FimGhostPrompt,
+	FimCompletionResult,
+	FillInAtCursorSuggestion,
+} from "../types"
+import { getProcessedSnippets } from "./getProcessedSnippets"
 import { getTemplateForModel } from "../../continuedev/core/autocomplete/templating/AutocompleteTemplate"
 import { GhostModel } from "../GhostModel"
-import { FillInAtCursorSuggestion } from "./HoleFiller"
 
-export interface FimGhostPrompt {
-	strategy: "fim"
-	autocompleteInput: AutocompleteInput
-	formattedPrefix: string
-	prunedSuffix: string
-}
-
-export interface FimCompletionResult extends ResponseMetaData {
-	suggestion: FillInAtCursorSuggestion
-}
+export type { FimGhostPrompt, FimCompletionResult }
 
 export class FimPromptBuilder {
 	constructor(private contextProvider: GhostContextProvider) {}
@@ -22,8 +18,14 @@ export class FimPromptBuilder {
 	 * Build complete FIM prompt with all necessary data
 	 */
 	async getFimPrompts(autocompleteInput: AutocompleteInput, modelName: string): Promise<FimGhostPrompt> {
-		const { filepathUri, helper, snippetsWithUris, workspaceDirs } =
-			await this.contextProvider.getProcessedSnippets(autocompleteInput, autocompleteInput.filepath)
+		const { filepathUri, helper, snippetsWithUris, workspaceDirs } = await getProcessedSnippets(
+			autocompleteInput,
+			autocompleteInput.filepath,
+			this.contextProvider.contextService,
+			this.contextProvider.model,
+			this.contextProvider.ide,
+			this.contextProvider.ignoreController,
+		)
 
 		// Use pruned prefix/suffix from HelperVars (token-limited based on DEFAULT_AUTOCOMPLETE_OPTS)
 		const prunedPrefixRaw = helper.prunedPrefix
