+++ src/services/ghost/types.ts
@@ -11,9 +11,15 @@ import { ContextRetrievalService } from "../continuedev/core/autocomplete/contex
 import { VsCodeIde } from "../continuedev/core/vscode-test-harness/src/VSCodeIde"
 import { GhostModel } from "./GhostModel"
 
-/**
- * Metadata returned from LLM API responses including cost and token usage
- */
+export const AUTOCOMPLETE_PROVIDER_MODELS = new Map([
+	["mistral", "codestral-latest"],
+	["kilocode", "mistralai/codestral-2508"],
+	["openrouter", "mistralai/codestral-2508"],
+	["bedrock", "mistral.codestral-2508-v1:0"],
+] as const)
+
+export type AutocompleteProviderKey = typeof AUTOCOMPLETE_PROVIDER_MODELS extends Map<infer K, any> ? K : never
+
 export interface ResponseMetaData {
 	cost: number
 	inputTokens: number
@@ -25,13 +31,10 @@ export interface ResponseMetaData {
 export interface GhostSuggestionContext {
 	document: vscode.TextDocument
 	range?: vscode.Range | vscode.Selection
-	recentlyVisitedRanges?: AutocompleteCodeSnippet[] // Recently visited code snippets for context
-	recentlyEditedRanges?: RecentlyEditedRange[] // Recently edited ranges for context
+	recentlyVisitedRanges?: AutocompleteCodeSnippet[]
+	recentlyEditedRanges?: RecentlyEditedRange[]
 }
 
-/**
- * Ghost-specific extensions to TabAutocompleteOptions
- */
 export interface GhostTabAutocompleteExtensions {
 	template?: string
 	useOtherFiles?: boolean
@@ -39,32 +42,16 @@ export interface GhostTabAutocompleteExtensions {
 	maxSnippetTokens?: number
 }
 
-/**
- * Tab autocomplete options for Ghost
- * Based on CoreTabAutocompleteOptions with some fields made optional and ghost-specific extensions
- */
 export type TabAutocompleteOptions = Partial<CoreTabAutocompleteOptions> & GhostTabAutocompleteExtensions
 
-/**
- * Recently edited range with timestamp
- * Duplicated from continuedev/core to avoid coupling
- */
 export interface RecentlyEditedRange extends RangeInFile {
 	timestamp: number
 	lines: string[]
 	symbols: Set<string>
 }
 
-/**
- * Code snippet for autocomplete context
- * Re-exported from continuedev/core for compatibility
- */
 export type { AutocompleteCodeSnippet }
 
-/**
- * Input for autocomplete request (CompletionProvider-compatible)
- * Duplicated from continuedev/core to avoid coupling
- */
 export interface AutocompleteInput {
 	isUntitledFile: boolean
 	completionId: string
@@ -81,10 +68,6 @@ export interface AutocompleteInput {
 	injectDetails?: string
 }
 
-/**
- * Output from autocomplete request (CompletionProvider-compatible)
- * Duplicated from continuedev/core to avoid coupling
- */
 export interface AutocompleteOutcome extends TabAutocompleteOptions {
 	accepted?: boolean
 	time: number
@@ -106,10 +89,6 @@ export interface AutocompleteOutcome extends TabAutocompleteOptions {
 	profileType?: "local" | "platform" | "control-plane"
 }
 
-/**
- * Result from prompt generation including prefix/suffix
- * New interface for Ghost to align with CompletionProvider
- */
 export interface PromptResult {
 	systemPrompt: string
 	userPrompt: string
@@ -118,13 +97,73 @@ export interface PromptResult {
 	completionId: string
 }
 
-// ============================================================================
-// Conversion Utilities
-// ============================================================================
+export interface FillInAtCursorSuggestion {
+	text: string
+	prefix: string
+	suffix: string
+}
 
-/**
- * Extract prefix and suffix from a document at a given position
- */
+export interface MatchingSuggestionResult {
+	text: string
+	matchType: CacheMatchType
+}
+
+export interface LLMRetrievalResult extends ResponseMetaData {
+	suggestion: FillInAtCursorSuggestion
+}
+
+export interface ChatCompletionResult extends ResponseMetaData {
+	suggestion: FillInAtCursorSuggestion
+}
+
+export interface FimCompletionResult extends ResponseMetaData {
+	suggestion: FillInAtCursorSuggestion
+}
+
+export interface FimGhostPrompt {
+	strategy: "fim"
+	autocompleteInput: AutocompleteInput
+	formattedPrefix: string
+	prunedSuffix: string
+}
+
+export interface HoleFillerGhostPrompt {
+	strategy: "hole_filler"
+	autocompleteInput: AutocompleteInput
+	systemPrompt: string
+	userPrompt: string
+}
+
+export type GhostPrompt = FimGhostPrompt | HoleFillerGhostPrompt
+
+export interface GhostStatusBarStateProps {
+	enabled?: boolean
+	model?: string
+	provider?: string
+	profileName?: string | null
+	hasValidToken: boolean
+	totalSessionCost: number
+	completionCount: number
+	sessionStartTime: number
+}
+
+export interface AutocompleteContext {
+	languageId: string
+	modelId?: string
+	provider?: string
+	strategy?: "fim" | "hole_filler"
+}
+
+export type CacheMatchType = "exact" | "partial_typing" | "backward_deletion"
+
+export type CostTrackingCallback = (cost: number, inputTokens: number, outputTokens: number) => void
+
+export interface PendingRequest {
+	prefix: string
+	suffix: string
+	promise: Promise<void>
+}
+
 export function extractPrefixSuffix(
 	document: vscode.TextDocument,
 	position: vscode.Position,
@@ -138,9 +177,6 @@ export function extractPrefixSuffix(
 	}
 }
 
-/**
- * Convert GhostSuggestionContext to AutocompleteInput
- */
 export function contextToAutocompleteInput(context: GhostSuggestionContext): AutocompleteInput {
 	const position = context.range?.start ?? context.document.positionAt(0)
 	const { prefix, suffix } = extractPrefixSuffix(context.document, position)
